refactor: change structure for controlled form
