+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export { UncontrolledForm } from './UncontrolledForm/UncontrolledForm';
-export { ControlledForm } from './ControlledForm/ControlledForm';
 export { DataFormList } from './DataFormList/DataFormList';
 export { DataForm } from './DataForm/DataForm';
