+++ src/features/controlledForm/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useControlledForm } from './useControlledForm';
