+++ src/features/controlledForm/index.ts
@@ -1,0 +1,1 @@
+export { ControlledForm } from './components/ControlledForm';
